JS: Refactor embed (#9653)
